Fix running with TChains & cloning output trees.
The problem with running over TChains was an initialization issue with
TClonesArrays.  This is now fixed.

Prior to this, cloning output trees did not work with TChains.
This was because of 2 problems: The first is that a new DTreeInterface
was being created for each new TTree, but the output tree pointer was
not being copied (oops). Now, the same DTreeInterface object is reused,
and the branch addresses are reinitialized to those in the new tree.

However, the nastier problem was that PROOF does not pass the TChain
into the selector; it only passes the current TTree.  Therefore, you
cannot use TChain::CloneTree(), because you can't get to the chain, and
you can't use TTree::CloneTree(), because the branch address pointers
change between trees when processed this way.  So, the new tree is
created manually.

Also, some bug fixes were made to how arrays were handled.